Enhance Zip file handling

+++ src/SSWR/AVIRead/AVIRPackageForm.cpp
@@ -1,4 +1,5 @@
 #include "Stdafx.h"
+#include "Crypto/Hash/HashCreator.h"
 #include "IO/Path.h"
 #include "IO/VirtualPackageFile.h"
 #include "SSWR/AVIRead/AVIRPackageForm.h"
@@ -27,8 +28,19 @@ typedef enum
 	MNU_DELETE_ITEM,
 	MNU_OPEN_TEXT_VIEWER,
 	MNU_OPEN_HEX_VIEWER,
-	MNU_APPEND_ZIP
+	MNU_APPEND_ZIP,
+	MNU_TEST
 } MenuItem;
+
+void AVIRPackageForm_TestHandler(const UInt8 *buff, UOSInt buffSize, void *userData)
+{
+	SSWR::AVIRead::AVIRPackageForm::ReadSession *sess = (SSWR::AVIRead::AVIRPackageForm::ReadSession*)userData;
+	if (sess->hash)
+	{
+		sess->hash->Calc(buff, buffSize);
+	}
+	sess->fileReadSize += buffSize;
+}
 
 UInt32 __stdcall SSWR::AVIRead::AVIRPackageForm::ProcessThread(void *userObj)
 {
@@ -452,6 +464,179 @@ void __stdcall SSWR::AVIRead::AVIRPackageForm::OnFiles(void *userObj, NotNullPtr
 	me->PasteFiles(fileNames, false);
 }
 
+void SSWR::AVIRead::AVIRPackageForm::TestPackage(NotNullPtr<IO::ActiveStreamReader> reader, NotNullPtr<ReadSession> sess, NotNullPtr<IO::PackageFile> pack)
+{
+	UTF8Char sbuff[512];
+	UTF8Char *sptr;
+	NotNullPtr<IO::PackageFile> innerPack;
+	NotNullPtr<IO::StreamData> stmData;
+	UInt64 fileSize;
+	UInt64 storeSize;
+	Bool needDelete;
+	Data::Timestamp startTime;
+	Data::Timestamp endTime;
+	UOSInt i = 0;
+	UOSInt j = pack->GetCount();
+	while (i < j)
+	{
+		switch (pack->GetItemType(i))
+		{
+		case IO::PackageFile::PackObjectType::PackageFileType:
+			sess->dirCnt++;
+			if (innerPack.Set(pack->GetItemPack(i, needDelete)))
+			{
+				TestPackage(reader, sess, innerPack);
+				if (needDelete)
+				{
+					innerPack.Delete();
+				}
+			}
+			else
+			{
+				sess->errorCnt++;
+				sess->sbError->Append(pack->GetSourceNameObj());
+				sess->sbError->AppendUTF8Char(IO::Path::PATH_SEPERATOR);
+				sptr = pack->GetItemName(sbuff, i);
+				if (sptr)
+				{
+					sess->sbError->AppendP(sbuff, sptr);
+				}
+				else
+				{
+					sess->sbError->Append(CSTR("file["));
+					sess->sbError->AppendUOSInt(i);
+					sess->sbError->AppendUTF8Char(']');
+				}
+				sess->sbError->Append(CSTR(" Error in getting directory object\r\n"));
+				break;
+			}
+			break;
+		case IO::PackageFile::PackObjectType::StreamData:
+			sess->fileCnt++;
+			if (stmData.Set(pack->GetItemStmDataNew(i)))
+			{
+				storeSize = pack->GetItemStoreSize(i);
+				fileSize = stmData->GetDataSize();
+				sess->fileReadSize = 0;
+				sess->totalSize += fileSize;
+				sess->totalStoreSize += storeSize;
+				if (pack->GetItemComp(i) == Data::Compress::Decompressor::CM_UNCOMPRESSED)
+				{
+					sess->totalSizeUncomp += fileSize;
+					if (sess->hash) sess->hash->Clear();
+					startTime = Data::Timestamp::UtcNow();
+					reader->ReadStreamData(stmData, 0);
+					endTime = Data::Timestamp::UtcNow();
+					if (sess->fileReadSize == fileSize)
+					{
+						Data::Duration thisTime = endTime.Diff(startTime);
+						sess->totalDurUncomp += thisTime;
+						Double thisSpeed = UInt64_Double(fileSize) / thisTime.GetTotalSec();
+						if (sess->slowestSpeedUncomp == 0 || sess->slowestSpeedUncomp > thisSpeed)
+						{
+							sess->slowestSpeedUncomp = thisSpeed;
+						}
+					}
+					else
+					{
+						sess->errorCnt++;
+						sess->sbError->Append(pack->GetSourceNameObj());
+						sess->sbError->AppendUTF8Char(IO::Path::PATH_SEPERATOR);
+						sptr = pack->GetItemName(sbuff, i);
+						if (sptr)
+						{
+							sess->sbError->AppendP(sbuff, sptr);
+						}
+						else
+						{
+							sess->sbError->Append(CSTR("file["));
+							sess->sbError->AppendUOSInt(i);
+							sess->sbError->AppendUTF8Char(']');
+						}
+						sess->sbError->Append(CSTR(" read size different from file size\r\n"));
+					}
+				}
+				else
+				{
+					sess->totalCompSize += fileSize;
+					if (sess->hash) sess->hash->Clear();
+					startTime = Data::Timestamp::UtcNow();
+					reader->ReadStreamData(stmData, 0);
+					endTime = Data::Timestamp::UtcNow();
+					if (sess->fileReadSize == fileSize)
+					{
+						Data::Duration thisTime = endTime.Diff(startTime);
+						sess->totalDurComp += thisTime;
+						Double thisSpeed = UInt64_Double(fileSize) / thisTime.GetTotalSec();
+						if (sess->slowestSpeedComp == 0 || sess->slowestSpeedComp > thisSpeed)
+						{
+							sess->slowestSpeedComp = thisSpeed;
+						}
+					}
+					else
+					{
+						sess->errorCnt++;
+						sess->sbError->Append(pack->GetSourceNameObj());
+						sess->sbError->AppendUTF8Char(IO::Path::PATH_SEPERATOR);
+						sptr = pack->GetItemName(sbuff, i);
+						if (sptr)
+						{
+							sess->sbError->AppendP(sbuff, sptr);
+						}
+						else
+						{
+							sess->sbError->Append(CSTR("file["));
+							sess->sbError->AppendUOSInt(i);
+							sess->sbError->AppendUTF8Char(']');
+						}
+						sess->sbError->Append(CSTR(" read size different from file size\r\n"));
+					}
+				}
+				endTime = Data::Timestamp::UtcNow();
+				stmData.Delete();
+			}
+			else
+			{
+				sess->errorCnt++;
+				sess->sbError->Append(pack->GetSourceNameObj());
+				sess->sbError->AppendUTF8Char(IO::Path::PATH_SEPERATOR);
+				sptr = pack->GetItemName(sbuff, i);
+				if (sptr)
+				{
+					sess->sbError->AppendP(sbuff, sptr);
+				}
+				else
+				{
+					sess->sbError->Append(CSTR("file["));
+					sess->sbError->AppendUOSInt(i);
+					sess->sbError->AppendUTF8Char(']');
+				}
+				sess->sbError->Append(CSTR(" Error in opening file for reading\r\n"));
+			}
+			break;
+		case IO::PackageFile::PackObjectType::ParsedObject:
+		case IO::PackageFile::PackObjectType::Unknown:
+			sess->errorCnt++;
+			sess->sbError->Append(pack->GetSourceNameObj());
+			sess->sbError->AppendUTF8Char(IO::Path::PATH_SEPERATOR);
+			sptr = pack->GetItemName(sbuff, i);
+			if (sptr)
+			{
+				sess->sbError->AppendP(sbuff, sptr);
+			}
+			else
+			{
+				sess->sbError->Append(CSTR("file["));
+				sess->sbError->AppendUOSInt(i);
+				sess->sbError->AppendUTF8Char(']');
+			}
+			sess->sbError->Append(CSTR(" file type is not supported\r\n"));
+			break;
+		}
+		i++;
+	}
+}
+
 void SSWR::AVIRead::AVIRPackageForm::DisplayPackFile(NotNullPtr<IO::PackageFile> packFile)
 {
 	UTF8Char sbuff[512];
@@ -699,10 +884,13 @@ SSWR::AVIRead::AVIRPackageForm::AVIRPackageForm(UI::GUIClientControl *parent, No
 	mnu = mnuMain->AddSubMenu(CSTR("&File"));
 	mnu->AddItem(CSTR("Save &As..."), MNU_SAVEAS, UI::GUIMenu::KM_CONTROL, UI::GUIControl::GK_S);
 	mnu->AddItem(CSTR("Append Zip Content"), MNU_APPEND_ZIP, UI::GUIMenu::KM_NONE, UI::GUIControl::GK_NONE);
+	mnu->AddItem(CSTR("Test from this folder"), MNU_TEST, UI::GUIMenu::KM_NONE, UI::GUIControl::GK_NONE);
 	mnu = mnuMain->AddSubMenu(CSTR("&Edit"));
 	mnu->AddItem(CSTR("&Paste"), MNU_PASTE, UI::GUIMenu::KM_CONTROL, UI::GUIControl::GK_V);
 	mnu->AddItem(CSTR("Copy To..."), MNU_COPYTO, UI::GUIMenu::KM_NONE, UI::GUIControl::GK_NONE);
 	mnu->AddItem(CSTR("Copy All To..."), MNU_COPYALLTO, UI::GUIMenu::KM_NONE, UI::GUIControl::GK_NONE);
+	mnu->AddItem(CSTR("Delete Item"), MNU_DELETE_ITEM, UI::GUIMenu::KM_NONE, UI::GUIControl::GK_DELETE);
+
 
 	NEW_CLASS(this->mnuPopup, UI::GUIPopupMenu());
 	this->mnuPopup->AddItem(CSTR("Copy To..."), MNU_COPYTO, UI::GUIMenu::KM_NONE, UI::GUIControl::GK_NONE);
@@ -767,6 +955,11 @@ SSWR::AVIRead::AVIRPackageForm::AVIRPackageForm(UI::GUIClientControl *parent, No
 		NotNullPtr<IO::VirtualPackageFile>::ConvertFrom(this->packFile)->GetInfoText(sb);
 		this->txtInfo->SetText(sb.ToCString());
 	}
+
+	this->tpTest = this->tcMain->AddTabPage(CSTR("Test Result"));
+	NEW_CLASS(this->txtTest, UI::GUITextBox(ui, this->tpTest, CSTR(""), true));
+	this->txtTest->SetReadOnly(true);
+	this->txtTest->SetDockType(UI::GUIControl::DOCK_FILL);
 
 	this->SetMenu(mnuMain);
 	this->HandleDropFiles(OnFiles, this);
@@ -945,13 +1138,23 @@ void SSWR::AVIRead::AVIRPackageForm::EventMenuClicked(UInt16 cmdId)
 			UOSInt i = this->PackFileIndex(this->lvFiles->GetSelectedIndex());
 			if (i != INVALID_INDEX)
 			{
-				if (!this->packFile->DeleteItem(i))
+				UTF8Char sbuff[512];
+				UTF8Char *sptr;
+				sptr = this->packFile->GetItemName(sbuff, i);
+				Text::StringBuilderUTF8 sb;
+				sb.Append(CSTR("Are you sure to delete \""));
+				sb.AppendP(sbuff, sptr);
+				sb.Append(CSTR("\"?"));
+				if (UI::MessageDialog::ShowYesNoDialog(sb.ToCString(), CSTR("Package"), this))
 				{
-					UI::MessageDialog::ShowDialog(CSTR("Cannot delete item"), CSTR("Package"), this);
-				}
-				else
-				{
-					this->DisplayPackFile(packFile);
+					if (!this->packFile->DeleteItem(i))
+					{
+						UI::MessageDialog::ShowDialog(CSTR("Cannot delete item"), CSTR("Package"), this);
+					}
+					else
+					{
+						this->DisplayPackFile(packFile);
+					}
 				}
 			}
 		}
@@ -974,6 +1177,7 @@ void SSWR::AVIRead::AVIRPackageForm::EventMenuClicked(UInt16 cmdId)
 					{
 						UI::MessageDialog::ShowDialog(CSTR("Error occurs during append zip file"), CSTR("Package"), this);
 					}
+					zipPkg.Delete();
 					this->DisplayPackFile(this->packFile);
 				}
 				else
@@ -985,6 +1189,65 @@ void SSWR::AVIRead::AVIRPackageForm::EventMenuClicked(UInt16 cmdId)
 		else
 		{
 			UI::MessageDialog::ShowDialog(CSTR("This directory is not supported to append zip content"), CSTR("Package"), this);
+		}
+		break;
+	case MNU_TEST:
+		{
+			Text::StringBuilderUTF8 sbError;
+			ReadSession sess;
+			IO::ActiveStreamReader reader(AVIRPackageForm_TestHandler, &sess, 1048576);
+			sess.currHash = Crypto::Hash::HashType::CRC32R_IEEE;
+			sess.hash = Crypto::Hash::HashCreator().CreateHash(sess.currHash);
+			sess.fileReadSize = 0;
+			sess.fileCnt = 0;
+			sess.dirCnt = 0;
+			sess.errorCnt = 0;
+			sess.totalSizeUncomp = 0;
+			sess.totalCompSize = 0;
+			sess.totalSize = 0;
+			sess.totalStoreSize = 0;
+			sess.totalDurUncomp = 0;
+			sess.totalDurComp = 0;
+			sess.slowestSpeedUncomp = 0;
+			sess.slowestSpeedComp = 0;
+			sess.sbError = &sbError;
+			Data::Timestamp startTime = Data::Timestamp::Now();
+			TestPackage(reader, sess, this->packFile);
+			Data::Timestamp endTime = Data::Timestamp::Now();
+			Text::StringBuilderUTF8 sb;
+			if (sess.errorCnt == 0)
+			{
+				sb.Append(CSTR("Files Read ok\r\n"));
+			}
+			else
+			{
+				sb.Append(CSTR("Files Read error found: "))->AppendUOSInt(sess.errorCnt)->Append(CSTR("\r\n"));
+				sb.Append(sbError);
+			}
+			sb.Append(CSTR("\r\n"));
+			Data::Duration totalTime = endTime.Diff(startTime);
+			sb.Append(CSTR("Start Time = "))->AppendTS(startTime)->Append(CSTR("\r\n"));
+			sb.Append(CSTR("End Time = "))->AppendTS(endTime)->Append(CSTR("\r\n"));
+			sb.Append(CSTR("Total Time = "))->AppendDur(totalTime)->Append(CSTR("\r\n"));
+			sb.Append(CSTR("File Count = "))->AppendUOSInt(sess.fileCnt)->Append(CSTR("\r\n"));
+			sb.Append(CSTR("Dir Count = "))->AppendUOSInt(sess.dirCnt)->Append(CSTR("\r\n"));
+			sb.Append(CSTR("Total decompressed size = "))->AppendU64(sess.totalSize)->Append(CSTR("\r\n"));
+			sb.Append(CSTR("Total store size = "))->AppendU64(sess.totalStoreSize)->Append(CSTR("\r\n"));
+			sb.Append(CSTR("Average Rate = "))->AppendDouble(UInt64_Double(sess.totalSize) / totalTime.GetTotalSec())->Append(CSTR("\r\n"));
+			sb.Append(CSTR("Total size of Uncomp file = "))->AppendU64(sess.totalSizeUncomp)->Append(CSTR("\r\n"));
+			sb.Append(CSTR("Total size of Comp file = "))->AppendU64(sess.totalCompSize)->Append(CSTR("\r\n"));
+			if (!sess.totalDurUncomp.IsZero())
+			{
+				sb.Append(CSTR("Average Uncomp Rate = "))->AppendDouble(UInt64_Double(sess.totalSizeUncomp) / sess.totalDurUncomp.GetTotalSec())->Append(CSTR("\r\n"));
+				sb.Append(CSTR("Slowest Uncomp Rate = "))->AppendDouble(sess.slowestSpeedUncomp)->Append(CSTR("\r\n"));
+			}
+			if (!sess.totalDurComp.IsZero())
+			{
+				sb.Append(CSTR("Average Comp Rate = "))->AppendDouble(UInt64_Double(sess.totalCompSize) / sess.totalDurComp.GetTotalSec())->Append(CSTR("\r\n"));
+				sb.Append(CSTR("Slowest Comp Rate = "))->AppendDouble(sess.slowestSpeedComp)->Append(CSTR("\r\n"));
+			}
+			this->txtTest->SetText(sb.ToCString());
+			this->tcMain->SetSelectedPage(this->tpTest);
 		}
 		break;
 	}
